Inserting a Card

+++ models/card.go
@@ -1,7 +1,7 @@
 package models
 
 type Card struct {
-	Id       int    `json:"id"`
+	Id       int64  `json:"id"`
 	Name     string `json:"name"`
 	Finished bool   `json:"finished"`
 }
